requirements.txt
Flask
pandas
numpy
joblib
scikit-learn==1.6.1
pythainlp
lime
imbalanced-learn
python-dotenv
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib